requirements.txt
Flask==3.0.3
psycopg2-binary==2.9.9
python-dotenv==1.0.1
requests==2.32.3
gunicorn==21.2.0

